separate out blob utilities
